+++ recipe/emailpassword/recipe.go
@@ -86,7 +86,7 @@ func recipeInit(config *epmodels.TypeInput) supertokens.Recipe {
 	}
 }
 
-func getRecipeInstanceOrThrowError() (*Recipe, error) {
+func GetRecipeInstanceOrThrowError() (*Recipe, error) {
 	if singletonInstance != nil {
 		return singletonInstance, nil
 	}